Broke out Entity::User::Validator class.
Basic User entity now knows nothing about validation per se; all it
does is delegate methods `#valid?` and `#invalid?` to its `@validator`
instance variable.

RSpec: 697 examples, 0 failures, 1 pending
RuboCop: 227 files inspected, no offenses detected
