+++ app/entities/user.rb
@@ -1,5 +1,5 @@
 
-require_relative 'user/internals/name_validator'
+require_relative 'user/validator'
 
 # Namespace containing all application-defined entities.
 module Entity
@@ -9,16 +9,11 @@ module Entity
   # which encapsulates more specific entity-oriented responsibilities.
   class User
     extend Forwardable
-    include ActiveModel::Validations
 
     def_delegator :attributes, :[]
+    def_delegators :@validator, :valid?, :invalid?
 
     attr_reader :email, :name, :profile
-
-    # NOTE: No uniqueness validation without database access.
-    validates :name, presence: true, length: { minimum: 6 }
-    validate :validate_name
-    validates_email_format_of :email
 
     # Initialise a new `Entity::User` instance, optionally specifying values for
     # setting attributes.
@@ -35,21 +30,13 @@ module Entity
       @name = attributes[:name]
       @email = attributes[:email]
       @profile = attributes[:profile]
+      @validator = Validator.new self
     end
 
     # Returns a Hash containing this instance's attribute values.
     # @return Hash Attributes (:name, :email, :profile) with values, in a Hash.
     def attributes
       instance_values.symbolize_keys
-    end
-
-    private
-
-    # Validation method called by ActiveModel validation.
-    # Instantiates internal object which validates name attribute and adds
-    # ActiveModel errors if any validations fail.
-    def validate_name
-      Internals::NameValidator.new(name).validate.add_errors_to_model(self)
     end
   end # class Entity::User
 end
